api/pyproject.toml
[project]
name = "Binbot"
version = "2.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "sqlmodel>=0.0.16",
    "python-jose",
    "passlib",
    "requests>=2.28.1",
    "websocket-client>=1.5.0",
    "websockets>=12.0",
    "requests-html>=0.10.0",
    "py3cw>=0.0.39",
    "apscheduler>=3.6.3",
    "requests-cache>=1.2.0",
    "kafka-python>=2.0.2",
    "py4j>=0.10.9",
    "sqlalchemy>=2.0.30",
    "alembic",
    "psycopg2",
    "alembic-postgresql-enum",
    "python-dotenv",
    "psycopg",
    "pybit",
    "pydantic-settings>=2.10.1",
    "confluent-kafka>=2.11.1",
    "aiokafka>=0.12.0",
    "kucoin-universal-sdk>=1.3.0",
    "pybinbot",
]

[project.urls]
Documentation = "https://carkod.github.io/binbot/"
Repository = "https://github.com/carkod/binbot"


[tool.uv]
package = false

[tool.ruff.lint]
select = [
    "I",  # isort
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e.g. for HTTPException
    "UP038",
]

[tool.mypy]
implicit_optional = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
warn_redundant_casts = true
check_untyped_defs = true
allow_redefinition = true
show_error_codes = true
follow_imports = "silent"
warn_unused_ignores = true
implicit_reexport = true
explicit_package_bases = true
disable_error_code = ["call-arg", "attr-defined"]

[project.optional-dependencies]
dev = [
    "pytest-asyncio>=1.2.0",
    "ruff>=0.11.12",
    "pytest>=8.1.1",
    "mongomock>=4.1.2",
    "httpx",
    "types-requests",
    "coverage",
    "pytest-vcr>=1.0.2",
    "mypy>=1.16.0",
    "types-requests>=2.32.0.20250515",
]
